Move a few services unused in controllers to lib
