+++ app/services/api/v1x1.rb
@@ -7,7 +7,6 @@ module Api
       class CopyPortfolio                       < Api::V1x0::Catalog::CopyPortfolio; end
       class CopyPortfolioItem                   < Api::V1x0::Catalog::CopyPortfolioItem; end
       class CreateIcon                          < Api::V1x0::Catalog::CreateIcon; end
-      class CreateRequestBodyFrom               < Api::V1x0::Catalog::CreateRequestBodyFrom; end
       class CreateRequestForAppliedInventories  < Api::V1x0::Catalog::CreateRequestBodyFrom; end
       class DuplicateImage                      < Api::V1x0::Catalog::DuplicateImage; end
       class ImportServicePlans                  < Api::V1x0::Catalog::ImportServicePlans; end
@@ -33,13 +32,6 @@ module Api
 
     module ServiceOffering
       class AddToPortfolioItem < Api::V1x0::ServiceOffering::AddToPortfolioItem; end
-    end
-
-    module Tags
-      class CollectLocalOrderResources < Api::V1x0::Tags::CollectLocalOrderResources; end
-      module Topology
-        class RemoteInventory < Api::V1x0::Tags::Topology::RemoteInventory; end
-      end
     end
   end
 end
