+++ spec/lib/catalog/create_request_body_from_spec.rb
@@ -1,4 +1,4 @@
-describe Api::V1x0::Catalog::CreateRequestBodyFrom, :type => [:service, :current_forwardable, :topology, :sources] do
+describe Catalog::CreateRequestBodyFrom, :type => [:service, :current_forwardable, :topology, :sources] do
   let(:subject) { described_class.new(order, order_item, task) }
   let(:order) { create(:order) }
   let(:order_item) { create(:order_item_with_callback) }
@@ -6,8 +6,8 @@ describe Api::V1x0::Catalog::CreateRequestBodyFrom, :type => [:service, :current
 
   describe "#process" do
     let(:sanitize_service_instance) { instance_double(Catalog::OrderItemSanitizedParameters, :sanitized_parameters => {:a => 1}) }
-    let(:local_tag_service_instance) { instance_double(Api::V1x0::Tags::CollectLocalOrderResources, :tag_resources => ["a"]) }
-    let(:remote_tag_service_instance) { instance_double(Api::V1x0::Tags::Topology::RemoteInventory, :tag_resources => ["b"]) }
+    let(:local_tag_service_instance) { instance_double(Tags::CollectLocalOrderResources, :tag_resources => ["a"]) }
+    let(:remote_tag_service_instance) { instance_double(Tags::Topology::RemoteInventory, :tag_resources => ["b"]) }
     let(:source_response) do
       SourcesApiClient::Source.new(:name => 'the platform')
     end
@@ -15,9 +15,9 @@ describe Api::V1x0::Catalog::CreateRequestBodyFrom, :type => [:service, :current
     before do
       allow(Catalog::OrderItemSanitizedParameters).to receive(:new).and_return(sanitize_service_instance)
       allow(sanitize_service_instance).to receive(:process).and_return(sanitize_service_instance)
-      allow(Api::V1x0::Tags::CollectLocalOrderResources).to receive(:new).with(:order_id => order.id).and_return(local_tag_service_instance)
+      allow(Tags::CollectLocalOrderResources).to receive(:new).with(:order_id => order.id).and_return(local_tag_service_instance)
       allow(local_tag_service_instance).to receive(:process).and_return(local_tag_service_instance)
-      allow(Api::V1x0::Tags::Topology::RemoteInventory).to receive(:new).with(task).and_return(remote_tag_service_instance)
+      allow(Tags::Topology::RemoteInventory).to receive(:new).with(task).and_return(remote_tag_service_instance)
       allow(remote_tag_service_instance).to receive(:process).and_return(remote_tag_service_instance)
 
       stub_request(:get, sources_url("sources/#{order_item.portfolio_item.service_offering_source_ref}"))
